Support #35538 - Molecular Analysis Run view page bug and test coverage
- Moved the 2 hooks inside useMolecularAnalysRun to their own hooks.
- Updated imports.
